Goreleaser (#2)
This pull request introduces several significant changes to automate the
release process, improve the README documentation, and enhance the
versioning system. The most important changes include the addition of a
GitHub Actions workflow for GoReleaser, the creation of a GoReleaser
configuration file, updates to the README for better user guidance, and
modifications to the versioning functionality.

### Automation and CI/CD:

* Added a GitHub Actions workflow (`.github/workflows/release.yml`) to
automate the release process using GoReleaser. This workflow triggers on
pull requests and pushes with version tags, sets up Go, and runs
GoReleaser to create releases.
* Created a GoReleaser configuration file (`.goreleaser.yml`) specifying
project details, build targets, archive settings, and release
information to streamline the release process.

### Documentation:

* Updated the `README.md` file to provide comprehensive instructions on
configuring and using the AWS SSO ECR Docker Credential Helper,
including prerequisites, Docker configuration, and usage examples.

### Versioning:

* Added a new function `SetVersionInfo` in `cmd/root.go` to set version
information dynamically, and updated the `main.go` to use this function
for setting version details.
[[1]](diffhunk://#diff-ab967ab1a2f3a1b769106eeb7bfe892ef0e81d1d27811fa15be08e6749feee1fR14-R22)
[[2]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261R7-R14)
* Removed the `version` command from `cmd/version.go`, as version
information is now set dynamically and printed in the root command.

### Error Handling:

* Improved error message in `internal/ecr/get_auth_credentials.go` to
provide clearer information when no valid configuration is found for a
specified profile.