Update README (#4)
This pull request includes several updates to the `README.md` file to
improve clarity and accuracy. The changes include adding a note about
the project's affiliation, correcting grammatical errors, and updating
configuration instructions.

Documentation improvements:

*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5R5-R7):
Added a note indicating that the project is not affiliated with Amazon
Web Services and is inspired by the Amazon ECR Docker Credential Helper.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L73-R76):
Corrected grammatical errors in installation instructions and clarified
the context for running the executable on Windows.
[[1]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L73-R76)
[[2]](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L88-R93)
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L106-R126):
Updated the example `credHelpers` section to use the correct helper name
`sso-ecr-login`.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L106-R126):
Added an example AWS SSO profile configuration to illustrate the use of
AWS SSO credentials.